+++ packages/vsce-api/__tests__/__unit__/ResourceTypes.test.ts
@@ -28,6 +28,8 @@ describe("ResourceType tests", () => {
       "CICSBundle",
       "CICSTSQueue",
       "CICSSharedTSQueue",
+      "CICSManagedRegion",
+      "CICSRegion",
     ]);
   });