+++ packages/vsce-api/src/interfaces/resources/index.ts
@@ -27,3 +27,4 @@ export * from "./ITask";
 export * from "./ITransaction";
 export * from "./IURIMap";
 export * from "./IWebService";
+export * from "./IRegions";