+++ packages/vsce/CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to the "cics-extension-for-zowe" extension will be documente
 ## Recent Changes
 
 - BugFix: Fixed Resource Inspector action menu so that it does not display beyond the panel edges. [#572](https://github.com/zowe/cics-for-zowe-client/issues/572)
+- Enhancement: Added show region logs and show SIT parameters actions to resource inspector. [#571] (https://github.com/zowe/cics-for-zowe-client/pull/571)
+- Enhancement: Removed show attributes for regions and resources. [#571] (https://github.com/zowe/cics-for-zowe-client/pull/571)
 - Enhancement: Added new "Inspect Region " command to quickly view CICS region information. [#568](https://github.com/zowe/cics-for-zowe-client/issues/568)
 - Enhancement: Organised profile tree alphabetically. [#552](https://github.com/zowe/cics-for-zowe-client/issues/552)
 - Enhancement: Open resource documentation from error notification. [#537](https://github.com/zowe/cics-for-zowe-client/issues/537)